Add support for the lowRISC Ibex CPU core
Ibex is a small 32 bit design formerly known as zero-riscy. It supports
RV32[IE]CM?. For compliance testing a compiled simulation of Ibex can
be used. This simulation model provides a special memory-mapped device
which can be used to interact with the simulator to dump the signature
and to terminate the simulation.

This commit only adds RV32I support.